拆分 wallet-operations.js 为 operations/ 按域分文件
将 1892 行的 wallet-operations.js 按业务域拆为 6 个文件，删除原文件，
message-handler 改为从各域文件直接导入：

- operations/wallet.js   钱包/账户/导出/改密/重置
- operations/tokens.js   余额/代币
- operations/sites.js    授权站点
- operations/contacts.js 联系人
- operations/backup-sync.js  WebDAV 备份同步/冲突解决
- operations/mpc.js      MPC 钱包/设置/会话/审计（含原散落在 backup 段的 MPC 常量）

域间无 handle 互调、helper 均域内自洽，故拆分零跨文件耦合。
各域文件使用精确 import。tests/change-password.test.mjs 导入路径同步更新。
测试全绿（39 单测 + 6 审批回归），dispatch 测试的 ReferenceError 扫描确认
6 个新文件无遗漏 import。
